+++ source/config/i18n.js
@@ -3,10 +3,14 @@
  * @author vivaxy
  */
 
-export const ONE_OR_THE_OTHER = (...propNames) => {
-    return `Properties ${propNames.join(' and ')} cannot be simultaneously defined. Please use one or the other.`
+export const $ONE_OR_THE_OTHER = (...propNames) => {
+    return `Properties ${propNames.join(' and ')} cannot be defined both. Please use one or the other.`
 };
 
-export const MUST_DEFINE_ONE = (...propNames) => {
+export const $MUST_DEFINE_ONE = (...propNames) => {
     return `Properties ${propNames.join(' or ')} must be defined.`
+};
+
+export const $INVALID_CHILDREN_TYPE = (parentName, childName) => {
+    return `Component '${parentName}' should only have children of '${childName}' type.`
 };